Separate Logger database from the rest
