+++ tracker/src/main/java/com/ericafenyo/tracker/logger/LogEntity.kt
@@ -26,10 +26,12 @@ package com.ericafenyo.tracker.logger
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
+import org.threeten.bp.LocalDateTime
 
 @Entity(tableName = "logs")
 data class LogEntity(
   @PrimaryKey(autoGenerate = true) val id: Int = 0,
+  val fts: String = LocalDateTime.now().toString(),
   val ts: Double = (System.currentTimeMillis() / 1000).toDouble(),
   val level: String,
   val message: String
